requirements.txt
click==8.1.7
colorama==0.4.6
dlib @ file:///C:/Users/DELL/Downloads/dlib-19.24.1-cp311-cp311-win_amd64.whl#sha256=6f1a5ee167975d7952b28e0ce4495f1d9a77644761cf5720fb66d7c6188ae496
face-recognition==1.2.3
face-recognition-models==0.3.0
mysql-connector-python==8.3.0
numpy==1.26.4
opencv-python==4.8.1.78
pillow==10.2.0
